+++ R/chaining.R
@@ -98,8 +98,7 @@
 #' 
 #' \code{\%>\%} - The \pkg{magrittr} "then" chain operator imported by 
 #' \pkg{dplyr}.  Imported for convenience.  See 
-#' \url{http://www.rdocumentation.org/packages/magrittr/functions/magrittr}
-#' for details.
+#' \url{https://github.com/smbache/magrittr} for details.
 #' 
 #' @param lhs The value to be piped.
 #' @param rhs A function or expression.